+++ wire_weaver_core/src/ast/api.rs
@@ -9,7 +9,7 @@ pub struct ApiLevel {
 
 #[derive(Debug)]
 pub struct ApiItem {
-    pub id: u16,
+    pub id: u32,
     pub docs: Vec<String>,
     pub multiplicity: Multiplicity,
     pub kind: ApiItemKind,
